Change default for auth_url in login form
The previous default was hardcoded to a particular IP address.  The
new default is derived from the host and protocol serving the zwift
ui, which is more likely to be correct.

This should probably be made a configuration value instead in the
future, but I do not see any current config facilities.