Add handleXhr helper
There is a repeated logic pattern when handling responses from swift:
call some function on an error, success, notFound, or auth issue.
Classifying a response is always done using the same http status
criteria.  I've added a helper that takes the functions as input,
using defaults when a function is not provided.